updated to 2019 frc version

+++ src/main/java/edu/flash3388/flashlib/frc/robot/hid/FrcHidInterface.java
@@ -47,7 +47,7 @@ public class FrcHidInterface implements HidInterface {
             return false;
         }
 
-        return mDriverStation.getStickButton(hid, (byte)button + 1);
+        return mDriverStation.getStickButton(hid, button + 1);
     }
 
     @Override
